# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Task Manager Apis",
    "description": "Task Manager REST Apis developed by Gokul Suthar",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "scheme": "bearer",
      "bearerFormat": "JWT",
      "name": "authorization"
    }
  },
  "security": [
    {
       "JWT": []
    }
 ],
  "definitions": {
...
    "updateMeSchema": {
      "properties": {
        "email": {
          "type": "string",
          "example": "current password"
        },
        "name": {
          "type": "string",
          "example": "new password"
        }
      }
    },
...
  },
  "paths": {
...
    "/users/{id}": {
      "get": {
        "tags": ["Admin"],
        "description": "Admin route to get user data from it's {id}!",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "User whose data you want to read!"
          }
        ], 
        "responses": {
          "200": {
            "description": "Sucess!"
          }
        }
      },

      "patch": {
        "tags": ["Admin"],
        "description": "User profile data update route",
        "consumes" : ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "User who wants to login!"
          },
          {
            "name": "user",
            "in": "body",
            "description": "Users who wants to update their profile data including profile-img!",
            "schema": {
              "$ref": "#/definitions/updateMeSchema"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success!"
          }, 
          "404": {
            "description": "No document found with that ID!"
          }, 
          "500": {
            "description": "Internal server error!"
          }
        }
      }
    },
...
  }
}